fix the font size calculations. We deal with pixel sizes here, not pointsizes!
svn path=/trunk/kdelibs/; revision=135869

+++ khtml/rendering/font.h
@@ -49,7 +49,7 @@ namespace khtml
 	}
 
 	QString family;
-	float size;
+	short int size;
 	bool italic 		: 1;
 	bool smallCaps 		: 1;
 	unsigned int weight 		: 8;
@@ -72,10 +72,10 @@ namespace khtml
 	}
 
 	void update( QPaintDeviceMetrics *devMetrics ) const;
-	
-	void drawText( QPainter *p, int x, int y, QChar *str, int len, int width, 
+
+	void drawText( QPainter *p, int x, int y, QChar *str, int len, int width,
                 QPainter::TextDirection d, int from=-1, int to=-1, QColor bg=QColor() ) const;
-	
+
 	int width( QChar *str, int len ) const;
 	int width( QChar ch ) const;
